+++ cf/test/test_Field.py
@@ -6,6 +6,7 @@ import os
 import re
 import tempfile
 import unittest
+from importlib.util import find_spec
 
 import numpy
 import numpy as np
@@ -1155,6 +1156,9 @@ class FieldTest(unittest.TestCase):
         with self.assertRaises(ValueError):
             f.insert_dimension(1, "qwerty")
 
+    @unittest.skipUnless(
+        find_spec("matplotlib"), "matplotlib required but not installed"
+    )
     def test_Field_indices(self):
         f = cf.read(self.filename)[0]