+++ cf/test/test_regrid_featureType.py
@@ -11,7 +11,7 @@ import cf
 
 esmpy_imported = True
 try:
-    import esmpy
+    import esmpy  # noqa: F401
 except ImportError:
     esmpy_imported = False
 
@@ -27,11 +27,6 @@ methods = (
 atol = 2e-12
 rtol = 0
 
-meshloc = {
-    "face": esmpy.MeshLoc.ELEMENT,
-    "node": esmpy.MeshLoc.NODE,
-}
-
 
 def esmpy_regrid(coord_sys, method, src, dst, **kwargs):
     """Helper function that regrids one dimension of Field data using
@@ -44,6 +39,11 @@ def esmpy_regrid(coord_sys, method, src, dst, **kwargs):
         Regridded numpy masked array.
 
     """
+    meshloc = {
+        "face": esmpy.MeshLoc.ELEMENT,
+        "node": esmpy.MeshLoc.NODE,
+    }
+
     esmpy_regrid = cf.regrid.regrid(
         coord_sys,
         src,
@@ -159,7 +159,7 @@ class RegridFeatureTypeTest(unittest.TestCase):
         self.assertFalse(cf.regrid_logging())
 
         # Create some nice data
-        src = self.dst_featureType
+        src = self.dst_featureType.copy()
         src.del_construct("cellmethod0")
         src = src[:12]
         src[...] = 273 + np.arange(12)