+++ Changelog.rst
@@ -17,6 +17,18 @@ Version NEXTVERSION
   (https://github.com/NCAS-CMS/cf-python/issues/???)
 * New optional dependency: ``healpix>=2025.1``
 * Changed dependency: ``cfdm>=1.13.0.0, <1.13.1.0``
+  
+Version 3.18.2
+--------------
+
+**2025-10-16**
+
+* Python 3.9 support removed
+  (https://github.com/NCAS-CMS/cf-python/issues/896)
+* Allow regridding for very large grids. New keyword parameter to
+  `cf.Field.regrids` and `cf.Field.regridc`: ``dst_grid_partitions``
+  (https://github.com/NCAS-CMS/cf-python/issues/878)
+* Changed dependency: ``Python>=3.10.0``
 
 ----