+++ docs/source/releases.rst
@@ -16,6 +16,7 @@ Documentation for all versions of cf.
 **CF-1.12**
 -----------
 
+* `Version 3.18.2 <https://ncas-cms.github.io/cf-python-docs/3.18.1>`_ (2025-10-16)
 * `Version 3.18.1 <https://ncas-cms.github.io/cf-python-docs/3.18.1>`_ (2025-08-20)
 * `Version 3.18.0 <https://ncas-cms.github.io/cf-python-docs/3.18.0>`_ (2025-06-05)
 * `Version 3.17.0 <https://ncas-cms.github.io/cf-python-docs/3.17.0>`_ (2025-04-02)